backend/requirements.txt
gunicorn==20.1.0
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
Django==3.1.12
django-csvimport==2.16
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-jet==1.0.8
django-prefetch==1.2.3
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-extra-fields==3.1.1
djoser==2.1.0
html5lib==1.1
idna==2.10
importlib-metadata==1.6.0
iniconfig==1.1.1
json-table-schema==0.2.1
lxml==4.6.3
messytables==0.15.2
more-itertools==8.2.0
packaging==20.9
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary
py==1.10.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-magic==0.4.22
pytz==2021.1
requests==2.25.1
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
wcwidth==0.1.9
webencodings==0.5.1
XlsxWriter==1.4.4
xlrd==2.0.1
zipp==3.1.0